requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.15.0
keras==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
setuptools==68.0.0